fix: refactored into handlers
