+++ lifecycle-operator/webhooks/pod_mutator/handlers/kubernetes_handler.go
@@ -1,0 +1,11 @@
+package handlers
+
+import (
+	"context"
+
+	corev1 "k8s.io/api/core/v1"
+)
+
+type K8sHandler interface {
+	Handle(ctx context.Context, pod *corev1.Pod, namespace string) error
+}
