+++ lifecycle-operator/webhooks/pod_mutator/handlers/pod_annotation_handler_test.go
@@ -239,7 +239,7 @@ func Test_copyAnnotationsIfParentAnnotated(t *testing.T) {
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			a := &PodHandler{
+			a := &PodAnnotationHandler{
 				Client: tt.fields.Client,
 				Log:    tt.fields.Log,
 			}
